refactor: 통합테스트 인수테스트로 네이밍 변경
