+++ src/test/java/cart/acceptance/AcceptanceTest.java
@@ -1,4 +1,4 @@
-package cart.integration;
+package cart.acceptance;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
@@ -8,7 +8,7 @@ import org.springframework.test.annotation.DirtiesContext;
 
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class IntegrationTest {
+public class AcceptanceTest {
     @LocalServerPort
     private int port;
